Changed project structure + pluggable views
* add blueprints
* add pluggable view (listview, profileview )
