+++ .flaskenv
@@ -1,2 +1,2 @@
-FLASK_APP=app
+FLASK_APP=manage
 FLASK_ENV=development
