+++ manage.py
@@ -1,0 +1,5 @@
+from videoblog import app
+
+
+if __name__ == "__main__":
+    app.run()
